requirements.txt
flask
gunicorn
pillow
boto3
exifread
geopy
